Initial work basing public API on CRoaring.Net
Some basic operations are implemented and the tests from the CRoaring.Net
library are passing for those operations.
The goal is to get as much working as I can before delving into sparse arrays,
runs, and packing/optimizing and rely on the red/black tree SortedSet<ushort>
for now. It will be interesting to run some benchmarks and see how well
it performs in its current state and what the upsides/downsides are to using
SortedSet.